combine all tips scripts into one
- can help make things a bit more DRY
- remove unused tip-post-processing.py script
